main-valid: add generation tracking to AST
This gives each AST node a "generation" which essentially means at which
transformation number it was created.

This allows us to prioritise mutations of more recently inserted AST nodes,
which I suspect is useful when applying multiple transformations in a row
to larger programs, since it will focus mutations in a particular place
rather than scattering them around everywhere.

Whether it is really useful or not remains to be seen, but we should at
least have the infrastructure to do things like this.